refactor: migrate orderId type to uint216 for consistency and heap compatibility
- Replaced all orderId-related `uint256` and `bytes32` types with `uint216`
- Updated function signatures, mappings, and interfaces accordingly
- Integrated TaskQueueLib using `uint216` for scheduling via MinHeap
- Adjusted encoding/decoding and helper methods to ensure safe narrowing casts
- Fixed heap indexing bugs due to 1-based indexing of mapping
- Ensured tests and invariants reflect new type definitions

+++ src/PaymentProcessorStorage.sol
@@ -12,7 +12,7 @@ contract PaymentProcessorStorage is IPaymentProcessorStorage, Ownable {
      * @notice The next available unique invoice ID.
      * @dev Used to track and increment standalone or sub-invoice identifiers.
      */
-    uint256 private nextInvoiceId;
+    uint216 private nextInvoiceId;
 
     /**
      * @notice Tracks whether an address is authorized to perform restricted actions.
@@ -43,7 +43,7 @@ contract PaymentProcessorStorage is IPaymentProcessorStorage, Ownable {
     }
 
     /// @inheritdoc IPaymentProcessorStorage
-    function updateInvoiceId(uint256 by) external onlyAuthorized returns (uint256) {
+    function updateInvoiceId(uint216 by) external onlyAuthorized returns (uint216) {
         nextInvoiceId += by;
         return totalInvoiceCreated();
     }
@@ -79,12 +79,12 @@ contract PaymentProcessorStorage is IPaymentProcessorStorage, Ownable {
     }
 
     /// @inheritdoc IPaymentProcessorStorage
-    function getNextInvoiceId() external view returns (uint256) {
+    function getNextInvoiceId() external view returns (uint216) {
         return nextInvoiceId;
     }
 
     /// @inheritdoc IPaymentProcessorStorage
-    function totalInvoiceCreated() public view returns (uint256) {
+    function totalInvoiceCreated() public view returns (uint216) {
         return nextInvoiceId - 1;
     }
